+++ src/zakuro_version.h
@@ -8,6 +8,9 @@ class ZakuroVersion {
   static std::string GetClientName();
   static std::string GetLibwebrtcName();
   static std::string GetEnvironmentName();
+  static std::string GetVersion();
+  static std::string GetSoraCppSdkVersion();
+  static std::string GetWebRTCVersion();
 };
 
 #endif  // ZAKURO_VERSION_H_